+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/_params.py
@@ -60,6 +60,9 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
                                     action='store_true', validator=add_progress_callback)
     socket_timeout_type = CLIArgumentType(help='The socket timeout(secs), used by the service to regulate data flow.',
                                           type=int)
+    num_results_type = CLIArgumentType(
+        default=5000, help='Specifies the maximum number of results to return. Provide "*" to return all.',
+        validator=validate_storage_data_plane_list)
 
     sas_help = 'The permissions the SAS grants. Allowed values: {}. Do not use if a stored access policy is ' \
                'referenced with --id that specifies this value. Can be combined.'
@@ -205,9 +208,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
 
     with self.argument_context('storage blob list') as c:
         c.argument('include', validator=validate_included_datasets)
-        c.argument('num_results', default=5000,
-                   help='Specifies the maximum number of blobs to return. Provide "*" to return all.',
-                   validator=validate_storage_data_plane_list)
+        c.argument('num_results', arg_type=num_results_type)
 
     with self.argument_context('storage blob generate-sas') as c:
         from .completers import get_storage_acl_name_completion_list
@@ -387,9 +388,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
         c.ignore('blob_name', 'snapshot')
 
     with self.argument_context('storage container list') as c:
-        c.argument('num_results', default=5000,
-                   help='Specifies the maximum number of containers to return. Provide "*" to return all.',
-                   validator=validate_storage_data_plane_list)
+        c.argument('num_results', arg_type=num_results_type)
 
     with self.argument_context('storage container set-permission') as c:
         c.ignore('signed_identifiers')
@@ -454,9 +453,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
         c.argument('protocol', arg_type=get_enum_type(['http', 'https'], 'https'), help='Protocol to use.')
 
     with self.argument_context('storage share list') as c:
-        c.argument('num_results', default=5000,
-                   help='Specifies the maximum number of shares to return. Provide "*" to return all.',
-                   validator=validate_storage_data_plane_list)
+        c.argument('num_results', arg_type=num_results_type)
 
     with self.argument_context('storage share exists') as c:
         c.ignore('directory_name', 'file_name')
@@ -553,9 +550,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
         from .completers import dir_path_completer
         c.argument('directory_name', options_list=('--path', '-p'), help='The directory path within the file share.',
                    completer=dir_path_completer)
-        c.argument('num_results', default=5000,
-                   help='Specifies the maximum number of files and directories to return. Provide "*" to return all.',
-                   validator=validate_storage_data_plane_list)
+        c.argument('num_results', arg_type=num_results_type)
 
     with self.argument_context('storage file metadata show') as c:
         c.register_path_argument()