+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/_validators.py
@@ -326,7 +326,6 @@ def validate_blob_type(namespace):
 def validate_storage_data_plane_list(namespace):
     if namespace.num_results == '*':
         namespace.num_results = None
-        return
     else:
         namespace.num_results = int(namespace.num_results)